refactor(insights): consolidate trends click handler into one shared module
The bar-chart and line-chart click handlers were near-identical copies — same
deps shape, same dataset resolution, same persons-modal call. Merge them into
viz/handleTrendsChartClick.tsx with a single TrendsChartClickDeps type and one
test suite covering both.
